Inline all the versions into the default branch
This takes all the version-specific branches and inline them into the
default branch, as they are.

This also prepares a placeholder for the upcoming v1.7 version which has
beta versions since a few days ago.

Several things are now paths-triggered:

* The build system will only trigger if the related paths have been
  changed.
* Renovate will trigger specific version updates based on the paths too.
* Versions <1.2.x are not updated/built anymore (it's already the case
  from before).

The build system is also now based on buildx, to prepare for the ARM
builds coming soon.