Split Nomad version into a dedicated file
The version used by Docker image during the build is now an argument
that must be passed from outside.

The version is extracted into a dedicated file that Renovate will keep
up to date.
This will allow, in a subsequent change, to automatically create new
releases when the version of Nomad changes.