Replace Alpine 3.18 by Debian 12 as the base image
This replaces the base image from Alpine 3.18 to Debian 12.

The main driver for this change (not implemented yet here) is to easily
support arm64 images in the future.

Nomad relies on the gblic for several functionalities and there are no
good up to date glibc packages for Alpine.
Changing the underlying libc of the Alpine image is also a bit risky and
defeats a bit the purpopse of using this base image in the first place.

This should also help to embark other Nomad plugins, and/or
Docker-related plugins by having a more general base image. After all,
Nomad itself is most probably expecting to run not on Alpine-based *VMs*
in real scenarios.

This increases the size of the final Docker image a bit: in my test, the
image increases from 150 MB to 190 MB, but it's still relatively small
and Debian opens more possibilities in the future.